+++ test_orders.py
@@ -1,6 +1,5 @@
 import unittest
 from flask import json
-import resources
 from models import *
 from app import app
 class test_orders(unittest.TestCase):
